fix: web search tool implementation and tests
Fixed issues with web search tool implementation and tests. Improved mock handling for the dependency, updated test structure, and ensured all tests are passing with good code coverage.
